+++ ExpandWorldPrefabs/PrefabData.cs
@@ -22,8 +22,8 @@ public class Data
   public string biomes = "";
   [DefaultValue(0f)]
   public float minDistance = 0f;
-  [DefaultValue(1f)]
-  public float maxDistance = 1f;
+  [DefaultValue(1000f)]
+  public float maxDistance = 1000f;
   [DefaultValue(-10000f)]
   public float minAltitude = -10000f;
   [DefaultValue(10000f)]
@@ -55,7 +55,7 @@ public class Info
 {
   public int Prefab = 0;
   public float Weight = 1f;
-  public int Swap = 0;
+  public Dictionary<int, int> Swaps = [];
   public string Data = "";
   public string Command = "";
   public bool Day = true;